@@ -12,15 +12,14 @@
     "npm": "^10.3.0"
   },
   "dependencies": {
-    "@emotion/react": "^11.11.4",
-    "@emotion/styled": "^11.11.0",
+    "@emotion/react": "^11.13.0",
+    "@emotion/styled": "^11.13.0",
     "@mdx-js/react": "^2.3.0",
-    "@mui/material": "^5.15.14",
-    "gatsby": "^5.13.3",
+    "@mui/material": "^5.16.6",
+    "gatsby": "^5.13.7",
     "gatsby-plugin-image": "^3.13.1",
     "gatsby-plugin-manifest": "^5.13.1",
     "gatsby-plugin-mdx": "^5.13.1",
-    "gatsby-plugin-react-helmet": "^6.13.1",
     "gatsby-plugin-sharp": "^5.13.1",
     "gatsby-plugin-sitemap": "^6.13.1",
     "gatsby-plugin-webfonts": "^2.3.2",
@@ -33,10 +32,15 @@
     "gatsby-source-filesystem": "^5.13.1",
     "gatsby-transformer-sharp": "^5.13.1",
     "prismjs": "^1.29.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
-    "style-to-object": "^1.0.5"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "style-to-object": "^1.0.6"
+  },
+  "overrides": {
+    "gatsby": {
+      "axios": "^1.2.1",
+      "webpack-dev-middleware": "^5.3.4"
+    }
   },
   "scripts": {
     "develop": "gatsby develop -H 0.0.0.0",
@@ -48,13 +52,13 @@
     "test": "jest"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^6.4.2",
-    "@testing-library/react": "^14.2.2",
+    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/react": "^16.0.0",
     "babel-jest": "^29.7.0",
-    "babel-preset-gatsby": "^3.13.1",
+    "babel-preset-gatsby": "^3.13.2",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
-    "prettier": "^3.2.5"
+    "prettier": "^3.3.3"
   }
 }
