@@ -1,18 +1,18 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
-// import userEvent from '@testing-library/user-event';
+import { render } from '@testing-library/react';
+
 import VersionSelect from './version-select';
+import { versions } from '../../__mocks__/documentation';
 
 const onChange = () => {
   return true;
 };
 
 test('Displays the documentation versions', () => {
-  const versions = ['1.0', '2.0'];
   const { getByTestId } = render(
     <VersionSelect
       versions={versions}
-      selectedVersion={'2.0'}
+      selectedVersion={versions[0]}
       showLabel={true}
       onChange={onChange}
     />
