@@ -3,7 +3,6 @@ import { graphql } from 'gatsby';
 
 import Hero from '../components/hero';
 import Layout from '../components/layout';
-import Seo from '../components/seo';
 
 const NotFoundPage = ({ data, location }) => {
   return (
@@ -13,12 +12,13 @@ const NotFoundPage = ({ data, location }) => {
           What you are looking for is either no longer here or has moved.
         </Hero>
       </div>
-      <Seo title="404: Not Found" />
     </Layout>
   );
 };
 
 export default NotFoundPage;
+
+export const Head = () => <title>404: Not Found</title>;
 
 export const pageQuery = graphql`
   query {
