@@ -3,14 +3,12 @@ import { graphql } from 'gatsby';
 
 import Layout from '../components/layout';
 import Section from '../components/section';
-import Seo from '../components/seo';
 
 const NewsPostTemplate = ({ children, data, pageContext, location }) => {
   const date = new Date(pageContext?.frontmatter.date);
 
   return (
     <Layout data={data} location={location}>
-      <Seo title={pageContext.title} description={pageContext.description} />
       <article className="page-main content">
         <h3>News</h3>
         <Section
@@ -25,6 +23,13 @@ const NewsPostTemplate = ({ children, data, pageContext, location }) => {
 };
 
 export default NewsPostTemplate;
+
+export const Head = ({ pageContext }) => (
+  <>
+    <title>{pageContext.title}</title>
+    <meta name="description" content={pageContext.description} />
+  </>
+);
 
 export const pageQuery = graphql`
   query ($id: String!) {
