@@ -1,10 +1,8 @@
 import { getCurrentVersion, getDocsVersion } from './version-docs';
-import documentation from '../../__mocks__/documentation';
+import { allDocs, slug, versions } from '../../__mocks__/documentation';
 
 test('Gets documentation by version', () => {
-  getDocsVersion(documentation, '1.0');
-
-  expect(getDocsVersion(documentation, '1.0')).toEqual([
+  expect(getDocsVersion(allDocs, '1.0')).toEqual([
     {
       id: '41748935-af24-51c3-8d32-458375e4246c',
       excerpt:
@@ -24,7 +22,25 @@ test('Gets documentation by version', () => {
 });
 
 test('Gets version of current documentation', () => {
+  expect(getCurrentVersion('/documentation/1.0/components/', versions)).toEqual(
+    '1.0'
+  );
+  expect(getCurrentVersion('/documentation/2.0/components/', versions)).toEqual(
+    '2.0'
+  );
   expect(
-    getCurrentVersion('/documentation/2.0/components/', ['1.0', '2.0'])
-  ).toEqual('2.0');
+    getCurrentVersion('/documentation/stuff/components/', [
+      ...versions,
+      'stuff',
+    ])
+  ).toEqual('stuff');
+  expect(
+    getCurrentVersion('/documentation/3.4.1-rc3/components/', [
+      ...versions,
+      '3.4.1-rc3',
+    ])
+  ).toEqual('3.4.1-rc3');
+  expect(
+    getCurrentVersion('/documentation/3.4.1-rc3/components/', versions)
+  ).toEqual(undefined);
 });
